+++ plugin-parent/common/src/main/java/jscover/maven/JasmineWebDriverRunner.java
@@ -1,6 +1,7 @@
 package jscover.maven;
 
-import org.apache.maven.plugin.MojoExecutionException;
+import static java.lang.String.format;
+
 import org.apache.maven.plugin.MojoFailureException;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
@@ -12,10 +13,8 @@ import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
 
-import static java.lang.String.format;
-
 public class JasmineWebDriverRunner extends WebDriverRunnerBase {
-    public void waitForTestsToComplete(WebDriver webClient) throws MojoExecutionException {
+    public void waitForTestsToComplete(WebDriver webClient) {
         new WebDriverWait(webClient, Duration.ofSeconds(timeOutSeconds)).until(ExpectedConditions.presenceOfElementLocated(By.className("jasmine-duration")));
         new WebDriverWait(webClient, Duration.ofSeconds(timeOutSeconds)).until(ExpectedConditions.textToBePresentInElementLocated(By.className("jasmine-duration"), "finished"));
     }
@@ -31,7 +30,7 @@ public class JasmineWebDriverRunner extends WebDriverRunnerBase {
     }
 
     public List<String> getFailures(WebDriver webClient) {
-        List<String> failures = new ArrayList<String>();
+        List<String> failures = new ArrayList<>();
         List<WebElement> elements = webClient.findElements(By.className("jasmine-failed"));
         for (WebElement element : elements) {
             List<WebElement> descriptions = element.findElements(By.className("jasmine-description"));