+++ plugin-parent/report/src/main/java/jscover/maven/MergeMojo.java
@@ -4,8 +4,6 @@ package jscover.maven;
 import jscover.report.ConfigurationForReport;
 import jscover.report.Main;
 import org.apache.maven.plugin.AbstractMojo;
-import org.apache.maven.plugin.MojoExecutionException;
-import org.apache.maven.plugin.MojoFailureException;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
@@ -17,11 +15,11 @@ import java.util.List;
 @Mojo(name = "merge", defaultPhase = LifecyclePhase.POST_INTEGRATION_TEST, threadSafe = true)
 public class MergeMojo extends AbstractMojo {
     @Parameter(required = true)
-    protected final List<String> mergeDirStrings = new ArrayList<String>();
+    protected final List<String> mergeDirStrings = new ArrayList<>();
     @Parameter(required = true)
     private File destDir;
 
-    public void execute() throws MojoExecutionException, MojoFailureException {
+    public void execute() {
         Main main = new Main();
         ConfigurationForReport config = getConfigurationForReport();
         main.setConfig(config);
@@ -32,7 +30,7 @@ public class MergeMojo extends AbstractMojo {
     private ConfigurationForReport getConfigurationForReport() {
         ConfigurationForReport config = new ConfigurationForReport();
         config.setMergeDestDir(destDir);
-        List<File> mergeDirs = new ArrayList<File>();
+        List<File> mergeDirs = new ArrayList<>();
         for (String mergeDirString : mergeDirStrings) {
             mergeDirs.add(new File(mergeDirString));
         }